lessons/lesson-04-common-pitfalls/src/paicc-4/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcript-analytics"
version = "0.1.0"
description = "Transcript analytics tool for processing text files, extracting word frequencies, generating visualizations, and providing AI-powered analysis"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["transcript", "analytics", "visualization", "openai"]
authors = [
    { name = "AI Coding Course", email = "example@email.com" }
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
transcript-analytics = "transcript_analytics.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/transcript_analytics"]